Moving SignedUrl tests to own file
