+++ google-cloud-storage/src/test/java/com/google/cloud/storage/it/ITStorageTest.java
@@ -28,13 +28,10 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeFalse;
-import static org.junit.Assume.assumeTrue;
 
 import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.apache.ApacheHttpTransport;
-import com.google.api.client.util.DateTime;
 import com.google.api.gax.paging.Page;
-import com.google.auth.ServiceAccountSigner;
 import com.google.auth.http.HttpTransportFactory;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.Condition;
@@ -78,8 +75,6 @@ import com.google.cloud.storage.HmacKey.HmacKeyState;
 import com.google.cloud.storage.HttpMethod;
 import com.google.cloud.storage.Notification;
 import com.google.cloud.storage.NotificationInfo;
-import com.google.cloud.storage.PostPolicyV4;
-import com.google.cloud.storage.PostPolicyV4.PostFieldsV4;
 import com.google.cloud.storage.Rpo;
 import com.google.cloud.storage.ServiceAccount;
 import com.google.cloud.storage.Storage;
@@ -119,16 +114,12 @@ import io.grpc.auth.MoreCallCredentials;
 import io.grpc.stub.MetadataUtils;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.Method;
-import java.net.URL;
-import java.net.URLConnection;
 import java.nio.ByteBuffer;
-import java.nio.file.Files;
 import java.security.Key;
+import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -148,11 +139,6 @@ import java.util.logging.Logger;
 import java.util.stream.StreamSupport;
 import java.util.zip.GZIPInputStream;
 import javax.crypto.spec.SecretKeySpec;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.mime.MultipartEntityBuilder;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.junit.AfterClass;
@@ -515,14 +501,6 @@ public class ITStorageTest {
   }
 
   @Test
-  public void testGetBucketEmptyFields() {
-    Bucket remoteBucket = storage.get(BUCKET, Storage.BucketGetOption.fields());
-    assertEquals(BUCKET, remoteBucket.getName());
-    assertNull(remoteBucket.getCreateTime());
-    assertNull(remoteBucket.getSelfLink());
-  }
-
-  @Test
   public void testGetBucketLifecycleRules() {
     String lifecycleTestBucketName = RemoteStorageHelper.generateBucketName();
     storage.create(
@@ -536,11 +514,11 @@ public class ITStorageTest {
                             .setAge(1)
                             .setNumberOfNewerVersions(3)
                             .setIsLive(false)
-                            .setCreatedBefore(new DateTime(System.currentTimeMillis()))
+                            .setCreatedBeforeOffsetDateTime(OffsetDateTime.now())
                             .setMatchesStorageClass(ImmutableList.of(StorageClass.COLDLINE))
                             .setDaysSinceNoncurrentTime(30)
-                            .setNoncurrentTimeBefore(new DateTime(System.currentTimeMillis()))
-                            .setCustomTimeBefore(new DateTime(System.currentTimeMillis()))
+                            .setNoncurrentTimeBeforeOffsetDateTime(OffsetDateTime.now())
+                            .setCustomTimeBeforeOffsetDateTime(OffsetDateTime.now())
                             .setDaysSinceCustomTime(30)
                             .build())))
             .build());
@@ -554,14 +532,14 @@ public class ITStorageTest {
               .getActionType()
               .equals(LifecycleRule.SetStorageClassLifecycleAction.TYPE));
       assertEquals(3, lifecycleRule.getCondition().getNumberOfNewerVersions().intValue());
-      assertNotNull(lifecycleRule.getCondition().getCreatedBefore());
+      assertNotNull(lifecycleRule.getCondition().getCreatedBeforeOffsetDateTime());
       assertFalse(lifecycleRule.getCondition().getIsLive());
       assertEquals(1, lifecycleRule.getCondition().getAge().intValue());
       assertEquals(1, lifecycleRule.getCondition().getMatchesStorageClass().size());
       assertEquals(30, lifecycleRule.getCondition().getDaysSinceNoncurrentTime().intValue());
-      assertNotNull(lifecycleRule.getCondition().getNoncurrentTimeBefore());
+      assertNotNull(lifecycleRule.getCondition().getNoncurrentTimeBeforeOffsetDateTime());
       assertEquals(30, lifecycleRule.getCondition().getDaysSinceCustomTime().intValue());
-      assertNotNull(lifecycleRule.getCondition().getCustomTimeBefore());
+      assertNotNull(lifecycleRule.getCondition().getCustomTimeBeforeOffsetDateTime());
     } finally {
       storage.delete(lifecycleTestBucketName);
     }
@@ -712,6 +690,8 @@ public class ITStorageTest {
         Storage.BlobTargetOption.kmsKeyName(kmsKeyOneResourcePath));
   }
 
+  //Seems quite similar to the normal KMS test? I get the subtle difference but is it worth creating a new bucket?
+  // or could we move the kms stuff out and just create the bucket with a default key?
   @Test
   public void testCreateBlobWithDefaultKmsKeyName()
       throws ExecutionException, InterruptedException {
@@ -2059,148 +2039,7 @@ public class ITStorageTest {
     }
   }
 
-  @Test
-  public void testGetSignedUrl() throws IOException {
-    if (storage.getOptions().getCredentials() != null) {
-      assumeTrue(storage.getOptions().getCredentials() instanceof ServiceAccountSigner);
-    }
-    String blobName = "test-get-signed-url-blob/with/slashes/and?special=!#$&'()*+,:;=?@[]";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    Blob remoteBlob = storage.create(blob, BLOB_BYTE_CONTENT);
-    assertNotNull(remoteBlob);
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-      URL url = storage.signUrl(blob, 1, TimeUnit.HOURS, urlStyle);
-      URLConnection connection = url.openConnection();
-      byte[] readBytes = new byte[BLOB_BYTE_CONTENT.length];
-      try (InputStream responseStream = connection.getInputStream()) {
-        assertEquals(BLOB_BYTE_CONTENT.length, responseStream.read(readBytes));
-        assertArrayEquals(BLOB_BYTE_CONTENT, readBytes);
-      }
-    }
-  }
 
-  @Test
-  public void testGetV2SignedUrlWithAddlQueryParam() throws IOException {
-    if (storage.getOptions().getCredentials() != null) {
-      assumeTrue(storage.getOptions().getCredentials() instanceof ServiceAccountSigner);
-    }
-    String blobName = "test-get-v2-with-generation-param";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    Blob remoteBlob = storage.create(blob, BLOB_BYTE_CONTENT);
-    assertNotNull(remoteBlob);
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-      String generationStr = remoteBlob.getGeneration().toString();
-      URL url =
-          storage.signUrl(
-              blob,
-              1,
-              TimeUnit.HOURS,
-              urlStyle,
-              Storage.SignUrlOption.withV2Signature(),
-              Storage.SignUrlOption.withQueryParams(
-                  ImmutableMap.<String, String>of("generation", generationStr)));
-      // Finally, verify that the URL works and we can get the object as expected:
-      URLConnection connection = url.openConnection();
-      byte[] readBytes = new byte[BLOB_BYTE_CONTENT.length];
-      try (InputStream responseStream = connection.getInputStream()) {
-        assertEquals(BLOB_BYTE_CONTENT.length, responseStream.read(readBytes));
-        assertArrayEquals(BLOB_BYTE_CONTENT, readBytes);
-      }
-    }
-  }
-
-  // TODO(b/144304815): Remove this test once all conformance tests contain query param test cases.
-  @Test
-  public void testGetV4SignedUrlWithAddlQueryParam() throws IOException {
-    if (storage.getOptions().getCredentials() != null) {
-      assumeTrue(storage.getOptions().getCredentials() instanceof ServiceAccountSigner);
-    }
-    String blobName = "test-get-v4-with-generation-param";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    Blob remoteBlob = storage.create(blob, BLOB_BYTE_CONTENT);
-    assertNotNull(remoteBlob);
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-      String generationStr = remoteBlob.getGeneration().toString();
-      URL url =
-          storage.signUrl(
-              blob,
-              1,
-              TimeUnit.HOURS,
-              urlStyle,
-              Storage.SignUrlOption.withV4Signature(),
-              Storage.SignUrlOption.withQueryParams(
-                  ImmutableMap.<String, String>of("generation", generationStr)));
-      // Finally, verify that the URL works and we can get the object as expected:
-      URLConnection connection = url.openConnection();
-      byte[] readBytes = new byte[BLOB_BYTE_CONTENT.length];
-      try (InputStream responseStream = connection.getInputStream()) {
-        assertEquals(BLOB_BYTE_CONTENT.length, responseStream.read(readBytes));
-        assertArrayEquals(BLOB_BYTE_CONTENT, readBytes);
-      }
-    }
-  }
-
-  @Test
-  public void testPostSignedUrl() throws IOException {
-    if (storage.getOptions().getCredentials() != null) {
-      assumeTrue(storage.getOptions().getCredentials() instanceof ServiceAccountSigner);
-    }
-    String blobName = "test-post-signed-url-blob";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    assertNotNull(storage.create(blob));
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-
-      URL url =
-          storage.signUrl(
-              blob, 1, TimeUnit.HOURS, Storage.SignUrlOption.httpMethod(HttpMethod.POST), urlStyle);
-      URLConnection connection = url.openConnection();
-      connection.setDoOutput(true);
-      connection.connect();
-      Blob remoteBlob = storage.get(BUCKET, blobName);
-      assertNotNull(remoteBlob);
-      assertEquals(blob.getBucket(), remoteBlob.getBucket());
-      assertEquals(blob.getName(), remoteBlob.getName());
-    }
-  }
-
-  @Test
-  public void testV4SignedUrl() throws IOException {
-    if (storage.getOptions().getCredentials() != null) {
-      assumeTrue(storage.getOptions().getCredentials() instanceof ServiceAccountSigner);
-    }
-
-    String blobName = "test-get-signed-url-blob/with/slashes/and?special=!#$&'()*+,:;=?@[]";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    Blob remoteBlob = storage.create(blob, BLOB_BYTE_CONTENT);
-    assertNotNull(remoteBlob);
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-
-      URL url =
-          storage.signUrl(
-              blob, 1, TimeUnit.HOURS, Storage.SignUrlOption.withV4Signature(), urlStyle);
-      URLConnection connection = url.openConnection();
-      byte[] readBytes = new byte[BLOB_BYTE_CONTENT.length];
-      try (InputStream responseStream = connection.getInputStream()) {
-        assertEquals(BLOB_BYTE_CONTENT.length, responseStream.read(readBytes));
-        assertArrayEquals(BLOB_BYTE_CONTENT, readBytes);
-      }
-    }
-  }
 
   @Test
   public void testGetBlobs() {
@@ -3463,43 +3302,6 @@ public class ITStorageTest {
     }
   }
 
-  @Test
-  public void testUploadUsingSignedURL() throws Exception {
-    String blobName = "test-signed-url-upload";
-    BlobInfo blob = BlobInfo.newBuilder(BUCKET, blobName).build();
-    assertNotNull(storage.create(blob));
-    Map<String, String> extensionHeaders = new HashMap<>();
-    extensionHeaders.put("x-goog-resumable", "start");
-    for (Storage.SignUrlOption urlStyle :
-        Arrays.asList(
-            Storage.SignUrlOption.withPathStyle(),
-            Storage.SignUrlOption.withVirtualHostedStyle())) {
-      URL signUrl =
-          storage.signUrl(
-              blob,
-              1,
-              TimeUnit.HOURS,
-              Storage.SignUrlOption.httpMethod(HttpMethod.POST),
-              Storage.SignUrlOption.withExtHeaders(extensionHeaders),
-              urlStyle);
-      byte[] bytesArrayToUpload = BLOB_STRING_CONTENT.getBytes();
-      Storage unauthenticatedStorage = StorageOptions.getUnauthenticatedInstance().getService();
-      try (WriteChannel writer = unauthenticatedStorage.writer(signUrl)) {
-        writer.write(ByteBuffer.wrap(bytesArrayToUpload, 0, bytesArrayToUpload.length));
-      }
-
-      int lengthOfDownLoadBytes = -1;
-      BlobId blobId = BlobId.of(BUCKET, blobName);
-      Blob blobToRead = storage.get(blobId);
-      try (ReadChannel reader = blobToRead.reader()) {
-        ByteBuffer bytes = ByteBuffer.allocate(64 * 1024);
-        lengthOfDownLoadBytes = reader.read(bytes);
-      }
-
-      assertEquals(bytesArrayToUpload.length, lengthOfDownLoadBytes);
-      assertTrue(storage.delete(BUCKET, blobName));
-    }
-  }
 
   @Test
   public void testBucketLocationType() throws ExecutionException, InterruptedException {
@@ -3582,31 +3384,6 @@ public class ITStorageTest {
       RemoteStorageHelper.forceDelete(storage, logsBucket, 5, TimeUnit.SECONDS);
       RemoteStorageHelper.forceDelete(storage, loggingBucket, 5, TimeUnit.SECONDS);
     }
-  }
-
-  @Test
-  public void testSignedPostPolicyV4() throws Exception {
-    PostFieldsV4 fields = PostFieldsV4.newBuilder().setAcl("public-read").build();
-
-    PostPolicyV4 policy =
-        storage.generateSignedPostPolicyV4(
-            BlobInfo.newBuilder(BUCKET, "my-object").build(), 7, TimeUnit.DAYS, fields);
-
-    HttpClient client = HttpClientBuilder.create().build();
-    HttpPost request = new HttpPost(policy.getUrl());
-    MultipartEntityBuilder builder = MultipartEntityBuilder.create();
-
-    for (Map.Entry<String, String> entry : policy.getFields().entrySet()) {
-      builder.addTextBody(entry.getKey(), entry.getValue());
-    }
-    File file = File.createTempFile("temp", "file");
-    Files.write(file.toPath(), "hello world".getBytes());
-    builder.addBinaryBody(
-        "file", new FileInputStream(file), ContentType.APPLICATION_OCTET_STREAM, file.getName());
-    request.setEntity(builder.build());
-    client.execute(request);
-
-    assertEquals("hello world", new String(storage.get(BUCKET, "my-object").getContent()));
   }
 
   @Test
